feat: add theming
